backend/requirements.txt
fastapi
uvicorn
passlib
python-jose
pydantic
motor
smtplib
